refactor: reorganize Web Components, save in same directory
